add file existance validation and logging

+++ pkg/repositories/lesson_material.py
@@ -6,8 +6,14 @@ import datetime
 
 from db.postgres import engine
 from db.models import LessonMaterial
-
 from logger.logger import logger
+
+
+import logging
+logging.basicConfig(
+    level=logging.DEBUG,
+    format='%(asctime)s - %(levelname)s - %(message)s'
+)
 
 
 MATERIAL_STORAGE = Path(Path.cwd(), 'lesson_materials')
@@ -28,20 +34,26 @@ def save_file(lesson_id, file):
 
 
 def upload_file(lesson_material: LessonMaterial):
-    
-    with Session(bind=engine) as db:
-        db.add(lesson_material)
-        db.commit()
-        return lesson_material.id
+    try:    
+        with Session(bind=engine) as db:
+            db.add(lesson_material)
+            db.commit()
+            return lesson_material.id
+    except Exception as e:
+        return {"message": e.args}
 
 
 def get_all_materials(lesson_id):
-    with Session(bind=engine) as db:
-        db_materials = db.query(LessonMaterial).filter(
-            LessonMaterial.deleted_at == None,
-            LessonMaterial.lesson_id == lesson_id
-        ).all()
-        return db_materials
+    try:
+        with Session(bind=engine) as db:
+            db_materials = db.query(LessonMaterial).filter(
+                LessonMaterial.deleted_at == None,
+                LessonMaterial.lesson_id == lesson_id
+            ).all()
+
+            return db_materials
+    except Exception as e:
+        return {"message": e.args}
     
 
 def get_material_by_id(file_id):
@@ -50,20 +62,28 @@ def get_material_by_id(file_id):
             LessonMaterial.deleted_at == None,
             LessonMaterial.id == file_id
         ).first()
+        logger.info(f"query returned: {db_material}")
         if db_material is None:
-            logger.error(f"File {file_id} not found")
+            logger.error(f"File ID {file_id} not found")
             return None
         
         file_path = str(Path(MATERIAL_STORAGE, str(db_material.lesson_id), db_material.hashed_filename))
-
+        logger.info(f"rep layer->get_material_by_id->file_path={file_path}, filename={db_material.filename}")
+        # if not file_path.exists():
+        #     return None
         return (file_path, db_material.filename)
     
 
 def replace_file(file_path, file):
     try:
+        file_path = Path(file_path)
+        logger.info(f"repo layer->replace_file->file_path is Path={isinstance(file_path, Path)}")
+        if not file_path.exists():
+            return None
+        
         file_path.unlink()  # удаляем старый файл
-
         new_file_path = Path(file_path.parent, file.hashed_filename)
+        logger.info(f"repo layer->replace_file->new_file_path={new_file_path}")
         with open(new_file_path, 'wb') as f:
             shutil.copyfileobj(file.file, f)
         return new_file_path
